requirements.txt
anyio==4.13.0
beautifulsoup4==4.14.3
build==1.5.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
contourpy==1.3.3
curl_cffi==0.15.0
cycler==0.12.1
exchange_calendars==4.13.2
fonttools==4.62.1
frozendict==2.4.7
h11==0.16.0
hmmlearn==0.3.3
httpcore==1.0.9
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
joblib==1.5.3
kiwisolver==1.5.0
korean-lunar-calendar==0.3.1
markdown-it-py==4.0.0
matplotlib==3.10.9
mdurl==0.1.2
multitasking==0.0.13
numpy==2.4.4
optuna==4.4.0
packaging==26.2
pandas==3.0.2
pandas_market_calendars==5.3.2
peewee==4.0.5
pillow==12.2.0
pip-tools==7.5.3
platformdirs==4.9.6
pluggy==1.6.0
protobuf==7.34.1
pyarrow==24.0.0
pycparser==3.0
Pygments==2.20.0
pyluach==2.3.0
pyparsing==3.3.2
pyproject_hooks==1.2.0
pytest==9.0.3
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
rich==15.0.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
soupsieve==2.8.3
threadpoolctl==3.6.0
toolz==1.1.0
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
websockets==16.0
wheel==0.47.0
yfinance==1.3.0
